+++ src/de/open4me/depot/datenobj/rmi/Bestand.java
@@ -1,5 +1,6 @@
 package de.open4me.depot.datenobj.rmi;
 
+import java.math.BigDecimal;
 import java.rmi.RemoteException;
 import java.util.Date;
 
@@ -12,14 +13,14 @@ public interface Bestand extends DBObject
 	public void setKontoid(Integer name) throws RemoteException;
 	public String getWPid() throws RemoteException;
 	public void setWPid(String name) throws RemoteException;
-	public Double getAnzahl() throws RemoteException;
-	public void setAnzahl(Double name) throws RemoteException;
-	public Double getKurs() throws RemoteException;
-	public void setKurs(Double name) throws RemoteException;
+	public BigDecimal getAnzahl() throws RemoteException;
+	public void setAnzahl(BigDecimal name) throws RemoteException;
+	public BigDecimal getKurs() throws RemoteException;
+	public void setKurs(BigDecimal name) throws RemoteException;
 	public String getKursw() throws RemoteException;
 	public void setKursw(String name) throws RemoteException;
-	public Double getWert() throws RemoteException;
-	public void setWert(Double name) throws RemoteException;
+	public BigDecimal getWert() throws RemoteException;
+	public void setWert(BigDecimal name) throws RemoteException;
 	public String getWertw() throws RemoteException;
 	public void setWertw(String name) throws RemoteException;
 	public Date getDatum() throws RemoteException;
