+++ src/de/open4me/depot/datenobj/rmi/Kursevents.java
@@ -1,5 +1,6 @@
 package de.open4me.depot.datenobj.rmi;
 
+import java.math.BigDecimal;
 import java.rmi.RemoteException;
 import java.util.Date;
 
@@ -10,8 +11,8 @@ public interface  Kursevents extends DBObject {
 	public void setWpid(Integer name) throws RemoteException;
 	public String getRatio() throws RemoteException;
 	public void setRatio(String name) throws RemoteException;
-	public Double getValue() throws RemoteException;
-	public void setValue(Double name) throws RemoteException;
+	public BigDecimal getValue() throws RemoteException;
+	public void setValue(BigDecimal name) throws RemoteException;
 	public String getAktion() throws RemoteException;
 	public void setAktion(String name) throws RemoteException;
 	public Date getDatum() throws RemoteException;
